requirements.txt
asgiref==3.8.1
certifi==2024.7.4
Django==4.1.13
django-cors-headers==4.4.0
django-elasticsearch-dsl==7.4
djangorestframework==3.15.1
djongo==1.3.6
elasticsearch==7.17.0
elasticsearch-dsl==7.4.1
ijson==3.3.0
line_profiler==4.1.3
pymongo==3.12.1
python-dateutil==2.9.0.post0
pytz==2024.1
six==1.16.0
sqlparse==0.2.4
tzdata==2024.1
urllib3==1.26.19
